refactor: merge _sv_diagram.py into _sv2dot.py, fix __main__ guard
- Remove _sv_diagram.py (15 lines), merge load_sv_dot + list_sv_categories
  into _sv2dot.py as module-level functions
- Add _dots_dir() helper above consumers
- Fix __name__ guard: only call main() when sys.argv has arguments
  Allows %run _sv2dot.py from QMD without triggering CLI
- Update sv_modules.qmd: %run _sv2dot.py instead of _sv_diagram.py
