--- docs/_include/_sv_diagram.py
@@ -1,30 +1,0 @@
-"""
-SV Architecture Diagram Loader
-==============================
-Provides load_sv_dot(category) — loads DOT by category.
-Categories match sv/ subdirectories: constraints, composition,
-projectors, pipeline, xif, all.
-Generated by build.yml: sv2dot.py --all outputs to _include/sv_dots/*.dot
-"""
-
-import os
-import glob
-
-
-def _dots_dir() -> str:
-    d = os.path.join(os.path.dirname(os.path.abspath(__file__)), "sv_dots")
-    os.makedirs(d, exist_ok=True)
-    return d
-
-
-def list_sv_categories() -> list[str]:
-    pattern = os.path.join(_dots_dir(), "*.dot")
-    return sorted(os.path.splitext(os.path.basename(f))[0] for f in glob.glob(pattern))
-
-
-def load_sv_dot(category: str = "all") -> str:
-    dot_path = os.path.join(_dots_dir(), f"{category}.dot")
-    if os.path.exists(dot_path):
-        with open(dot_path) as f:
-            return f.read()
-    return ""
