+++ docs/build.yml
@@ -16,7 +16,7 @@ target_config:
 # Commands listed here run AFTER the defaults.
 pre_build:
   _global:
-    - python3 _include/sv2dot.py ../poc/baremetal_riscv/sv/ --all
+    - python3 _include/_sv2dot.py ../poc/baremetal_riscv/sv/ --all
 
 # Post-render commands (additional to built-in defaults)
 # Built-in defaults: generate_llms_txt
