+++ docs/poc/architecture/sv_modules.qmd
@@ -20,7 +20,7 @@ from pathlib import Path
 _include = Path("../../_include").resolve()
 if str(_include) not in sys.path:
     sys.path.insert(0, str(_include))
-%run ../../_include/_sv_diagram.py
+%run ../../_include/_sv2dot.py
 ```
 
 ## Constraints
